Add files via upload
Recommendation engines are a subclass of machine learning which generally deal with ranking or rating products / users. Loosely defined, it is a system which predicts ratings a user might give to a specific item. These predictions will then be ranked and returned back to the user.Here take a look at our Recommendor system which recommends us the courses available on udemy